Fixes integer type warnings
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/sphinxbase@12665 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ src/libsphinxbase/util/bio.c
@@ -598,7 +598,7 @@ bio_read_wavfile(char const *directory,
 {
     FILE *uttfp;
     char *inputfile;
-    int32 n, l;
+    size_t n, l;
     int16 *data;
 
     n = strlen(extension);
